requirements.txt
flask
python-dotenv
flask-cors
# For MongoDB connection & getting rid of hardcoded product data
pymongo